requirements.txt
Flask==3.1.0
requests==2.32.3
Werkzeug==3.1.3
python-docx==1.1.2
pandas==2.2.3
numpy==2.2.2
faiss-gpu==1.8.2
faiss-cpu==1.10.0
torch==2.6.0+cu118
langchain==0.3.21
langchain-community==0.3.20
langchain-huggingface==0.1.2
transformers==4.50.0
python-dotenv==1.0.1